Make sure shell permissions are dropped if a test fails.
If a test requires shell permissions, it should drop them
at the end of the test even in case of failure.
In this case, a failure in an unrelated test caused successive
tests to be executed with shell permissions when they should have
not.
Tests failed since they were expecting a security exception
due to the lack of permissions, but the tests ended up being
called with shell permissions so they had the permission needed
to access the APIs on DeviceLockManager.

Bug: 255180414
Bug: 255180437
Test: atest CtsDeviceLockTestCases
Change-Id: Ie60dd47efb15bcbb13277af5d9e1565049218565