Create a worker class to start lock task mode
This worker will retry itself until device is currently in lock task
mode. The worker is a unique work item and will not be replaced or
duplicated by multiple schdeuling. The work will be stopped proactively
when device exit lock task mode.

Also fixed some errors in the scheduling of the device check in work.

Bug: 265080063
Test: Built Successfully.
Test: atest DeviceLockControllerRoboTests
Change-Id: I03bd33f4ded112cd2bcc701c31cee78df18b78cb