Replace component allowlist from DlcLockedBootCompletedReceiver.
We currently have a list of components that should not be automatically
disabled for secondary users.
This list comprises all components (except activities) in the
Android manifest of the Device Lock Controller, since we can
explicitly disable (if needed) components for secondary users by specifying
systemUserOnly="true" in the manifest.
The only components that are actually disabled are the ones from third
party libraries.
We can therefore just check for the component class prefix to make
a decision if a components needs to be disabled or not.

Bug: 266139344
Test: atest DeviceLockControllerRoboTests

Change-Id: Ie96ce9a73eb62dab0b0772bbfb75d554acc37ec6