Forcefully stop lock task mode in DLC
Setting an empty allowlist actually removes the DLC's policy and will
signal the system to start using the next admin's policy. If the device
is in locked task mode when we do this and one of the locked tasks is in
the next admin's policy, the device will stay in locked task mode even
though we wanted to exit the mode.

By updating the list to be an allowlist of an empty string, we stay the
enforcing admin and enforce a policy with no allowed tasks. This closes
any locked tasks, effectively ending lock task mode. We then properly
send an empty allowlist to remove our policy so other admins can manage
the device.

Bug: 288886570
Test: repro case in bug. does not get stuck in lock task mode
Change-Id: Ieb29df5f8866ac0c59175f7963c0fd8b3e5b2fe6