Disable DLC when restrictions are cleared (1/3)
This CL introduces the classes that interface with the rest of the
architecture and the basic control flow.

The FinalizationController is the basic public interface here with an
API to be notified when the device has had its restrictions cleared. The
implementation is then responsible for kicking off the work to report
this to the server and disabling the package afterward. Some of this
existing logic is moved into FinalizationControllerImpl, and the
controller decides which to do based on what state its in (unfinalized,
reporting, finalized).

Later CL's will implement persisting the finalized state on device and
disbling the package for real.

Bug: 279517666
Test: atest DeviceLockControllerRoboTests
Test: kiosk app "Clear restrictions" button
Change-Id: I4a9177bdc0f070d79677377cb3a4d05310fbdd39