Disable DLC when restrictions are cleared (2/3)
This CL adds reading/writing the finalization state to disk to ensure we
know if we need to report / disable on a boot.

Finalization state is persisted to disk using shared preferences at the
global device level.

As the writing API returns a future, we refactor the serialized
execution mechanism to use an ExecutionSequencer instead of a
SequentialExecutor as an ExecutionSequencer enforces that the future
finishes first before executing the next callable.

Bug: 279517666
Test: atest DeviceLockControllerRoboTests
Change-Id: Ib895f9897fe4cb31c3f585c8f7fa637f5ece5c9d