Disable package when finalization is complete
Disable the DLC package once we have determined we should be finalized.

Because the package is disabled and will not run, the system service has
to also persist a finalized state to know not to re-enable the DLC on
boot. This is done by wring to a simple XML file in global device state
once the DLC has determined that everything is finished.

Bug: 279517666
Test: atest DeviceLockControllerRoboTests
Change-Id: I30c29b27586c13c84901ea8b81f0542bfa543b45