Deflake finalization controller test
Test was flaking because we constructed the controller twice so both
would attempt to write their initial state to disk. This caused a race
where the state would be variable right before the test in question.

We fix this by ensuring only one runs.

We also protect against other flakes by directly waiting on the
initialization future.

Bug: 279517666
Test: atest FinalizationControllerImplTest --iterations 10
Change-Id: Ie6d8a7a317dd8b84b11eb4b119fc3b460c8709da