Check isProvisionReady before check-in
Checks the global parameter, isProvisionReady, to determine whether a
check-in is necessary.

This prevents the edge case of check-in happening again before the
provision state has left the UNPROVISIONED state (e.g. if it is in
set-up wizard and we don't want enter the PROVISIONING_IN_PROGRESS state
yet).

Also remove an unnecessary profile check and replace with an explicit
assertion that it the receiver is only run by the system user.

Bug: 301343214
Test: atest CheckInBootCompletedReceiverTest
Change-Id: I35e57ec5244ea2fa20fd64f78a0abf4bccf2846c