[DLC vNext] Add logging for Kiosk app requests
Adding logging line to DeviceLockControllerService under the
DeviceLockController package. This means the pushed atom will be
triggered when the Kiosk app's request reaches the DLC apex.

This CL also continues to use the StatsLogger class added in the previous CL. Also adds unit tests for the change in the DeviceLockControllerService.

Test: atest DeviceLockControllerRoboTests
Bug: b/259302076, b/309038511
Change-Id: I89b170cb55c6fda89e26acc3a7041cff54546810