[DLC vNext] Add tests for the atoms for logging
Adds android_test Soong build target for testing the behavior of the
newly added StatsLoggerImpl class, verifying it calling the static write
method of the automatically generated Java class from the atoms. Use
"atest DeviceLockControllerAndroidTests" command to run the new tests.
And the test requires a device to be connected to the computer.

I did not adopt the "verify what's written to the StatsLog class in
proto" approach because I think it is testing against an implementation
of a auto-generated class that is out of our control.

Test: atest DeviceLockControllerAndroidTests
Bug: b/305984988, b/259302076
Change-Id: I4993c52eb938f3aeebcc07de2760cc2ff75f30d8