Move setup listener to system service side
The setup listener is removed if the process hosting it dies which
causes provisioning to not start sometimes. This can be resolved by
ensuring that the system service listens for the setup completion signal
and calls into the device lock process.

This CL adds a "onUserSetupComplete" method similar to "onUserUnlock"
which implementations can use to trigger provisioning behavior.

Bug: 304433532
Test: manual
Test: atest DeviceLockControllerRoboTests

Change-Id: If6f0dfc9ae0b5f7915806104109237bc15e571fb