[DLC vNext] Remove Logging for provision completed in check-in request
As discussed with zzhen@, since the underlying gRPC call has already
been removed for the CheckInRequestReported with the provision completed
state, we should remove the related logging lines too. This CL removes
the logging code for this. And the related enum in the protobuf
definition of the atom is marked as deprecated in ag/25426868.

Test: atest DeviceLockControllerRoboTests
Test: atest DeviceLockControllerAndroidTests
Bug: b/312562483, b/259302076
Change-Id: Icdd7d92bb149412a0a24895df9809043c5a7cfe9