Retry check-in w/ backoff for more error codes
Utilize WorkManager's backoff policy to retry a failed gRPC in one
minute with doubling exponential backoff up to a max of 5 hours (limit
imposed by WorkManager).

Also increase the error codes that we should retry for (i.e. anything
that we think might go away within 5 hours).

With these changes, an error code that is not recoverable is now very
unexpected and should not happen, so upgrade logs in the failure path
to ERROR level instead of WARN.

Apply these changes to all workers that communicate with the server.

Bug: 313953173
Test: atest DeviceLockControllerRoboTests
Test: repro in bug
Change-Id: I3ce68e0b8b83ed7cde8d99bc8afd3ac26bbde831