Fix background thread work running on main thread
IsDeviceInApprovedCountryWorker was throwing an exception becuase it
expects to only be run on a background thread but due to a
directExecutor usage, it could run on main thread.

This is fixed by making sure to use the background thread instead of the
direct executor. I've also changed this for other workers that do
non-trivial transforms.

Bug: 314376083
Test: repro in bug
Change-Id: I2d303ad87c7820a85c9eda0496d924cbee0d5020